requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
bokeh==3.7.2
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
docarray==0.41.0
durationpy==0.9
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-experimental==0.3.4
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langgraph==0.3.21
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.3.19
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.32.0
numexpr==2.10.2
numpy==2.2.2
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.69.0
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
panel==1.6.2
param==2.2.0
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
posthog==3.23.0
prettytable==3.14.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.0
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyviz_comms==3.0.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.2
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
